ai-service/requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
python-dotenv
